+++ src/timeline.ts
@@ -1,8 +1,9 @@
 import { compareAsc, compareDesc } from 'date-fns';
+import Table from 'cli-table3';
 import { EntryLine, TimelineEntry, TimelineOptions } from './interfaces';
 import { DateFormat, Theme } from './util';
 
-export class Timeline {
+export default class Timeline {
   dateFormat: DateFormat;
   defaultSortDir: string;
   header: EntryLine | undefined;
@@ -23,6 +24,34 @@ export class Timeline {
 
   deleteEntry(index: number) {
     this.entries.splice(index, 1);
+  }
+
+  render() {
+    const table = new Table({
+      head: this.header?.fields.map((ef) => this.theme.stylize(ef.value, 'secondary')),
+      chars: {
+        top: this.theme.stylize('─', 'fade'),
+        'top-mid': this.theme.stylize('┬', 'fade'),
+        'top-left': this.theme.stylize('┌', 'fade'),
+        'top-right': this.theme.stylize('┐', 'fade'),
+        bottom: this.theme.stylize('─', 'fade'),
+        'bottom-mid': this.theme.stylize('┴', 'fade'),
+        'bottom-left': this.theme.stylize('└', 'fade'),
+        'bottom-right': this.theme.stylize('┘', 'fade'),
+        left: this.theme.stylize('│', 'fade'),
+        'left-mid': this.theme.stylize('├', 'fade'),
+        mid: this.theme.stylize('─', 'fade'),
+        'mid-mid': this.theme.stylize('┼', 'fade'),
+        right: this.theme.stylize('│', 'fade'),
+        'right-mid': this.theme.stylize('┤', 'fade'),
+        middle: this.theme.stylize('│', 'fade'),
+      },
+      style: {
+        border: [],
+      }
+    });
+
+    console.log(table.toString());
   }
 
   sort(sorter: (a: TimelineEntry, b: TimelineEntry) => number = this.defaultSort) {